+++ src/main/java/gift/doamin/user/service/AuthService.java
@@ -2,36 +2,33 @@ package gift.doamin.user.service;
 
 import gift.doamin.user.dto.LoginRequest;
 import gift.doamin.user.dto.SignUpRequest;
-import gift.doamin.user.entity.RefreshToken;
 import gift.doamin.user.entity.User;
 import gift.doamin.user.entity.UserRole;
-import gift.doamin.user.exception.InvalidRefreshTokenException;
 import gift.doamin.user.exception.InvalidSignUpFormException;
 import gift.doamin.user.exception.UserNotFoundException;
-import gift.doamin.user.repository.RefreshTokenRepository;
 import gift.doamin.user.repository.UserRepository;
 import gift.global.util.JwtDto;
-import gift.global.util.JwtProvider;
 import java.util.Optional;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class AuthService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
-    private final JwtProvider jwtProvider;
-    private final RefreshTokenRepository refreshTokenRepository;
+    private final AuthTokenService authTokenService;
+
 
     public AuthService(UserRepository userRepository, PasswordEncoder passwordEncoder,
-        JwtProvider jwtProvider, RefreshTokenRepository refreshTokenRepository) {
+        AuthTokenService authTokenService) {
         this.userRepository = userRepository;
         this.passwordEncoder = passwordEncoder;
-        this.jwtProvider = jwtProvider;
-        this.refreshTokenRepository = refreshTokenRepository;
+        this.authTokenService = authTokenService;
     }
 
+    @Transactional
     public JwtDto signUp(SignUpRequest signUpRequest) {
         String email = signUpRequest.getEmail();
         if (userRepository.existsByEmail(email)) {
@@ -43,9 +40,10 @@ public class AuthService {
         User user = new User(email, password, null, UserRole.USER);
         user = userRepository.save(user);
 
-        return genrateToken(user);
+        return authTokenService.genrateToken(user);
     }
 
+    @Transactional
     public JwtDto login(LoginRequest loginRequest) {
         String email = loginRequest.getEmail();
         String password = loginRequest.getPassword();
@@ -55,30 +53,6 @@ public class AuthService {
             throw new UserNotFoundException();
         }
 
-        return genrateToken(user.get());
-    }
-
-    public String makeNewAccessToken(String refreshToken) {
-        if (!jwtProvider.validateToken(refreshToken)) {
-            throw new InvalidRefreshTokenException();
-        }
-
-        RefreshToken tokenEntity = refreshTokenRepository.findByToken(refreshToken)
-            .orElseThrow(InvalidRefreshTokenException::new);
-
-        User user = tokenEntity.getUser();
-        return jwtProvider.generateAccessToken(user.getId(), user.getRole());
-    }
-
-    private JwtDto genrateToken(User user) {
-        JwtDto jwt = jwtProvider.generateToken(user.getId(), user.getRole());
-
-        RefreshToken refreshToken = refreshTokenRepository.findByUser(user)
-            .orElseGet(() -> new RefreshToken(jwt.getRefreshToken(), user));
-
-        refreshToken.setToken(jwt.getRefreshToken());
-
-        refreshTokenRepository.save(refreshToken);
-        return jwt;
+        return authTokenService.genrateToken(user.get());
     }
 }
